+++ asyncapi/exceptions.py
@@ -28,3 +28,19 @@ class InvalidJsonSchemaTypeError(AsyncApiError):
 
 class InvalidMessageError(AsyncApiError):
     ...
+
+
+class EmptyServersError(AsyncApiError):
+    ...
+
+
+class ServerNotFoundError(AsyncApiError):
+    ...
+
+
+class InvalidContentTypeError(AsyncApiError):
+    ...
+
+
+class InvalidAsyncApiVersionError(AsyncApiError):
+    ...