+++ docs/src/auto_spec/publish_http.py
@@ -5,7 +5,7 @@ import asyncio
 from asyncapi import build_api
 
 
-api = build_api('http://localhost:5000/asyncapi.json')
+api = build_api('http://localhost:5000/asyncapi.yaml')
 channel_id = 'user/update'
 message = api.payload(channel_id, id='fake-user', name='Fake User', age=33)